Make customizing the retry behavior easier
