+++ libcloud/utils/misc.py
@@ -17,16 +17,15 @@ from typing import List
 
 import os
 import binascii
-import socket
-import time
-import ssl
-from datetime import datetime, timedelta
-from functools import wraps
 
-from libcloud.utils.py3 import httplib
-from libcloud.common.exceptions import RateLimitReachedError
 from libcloud.common.providers import get_driver as _get_driver
 from libcloud.common.providers import set_driver as _set_driver
+# Imported for backward compatibility
+# noinspection PyProtectedMember
+from libcloud.utils.retry import (Retry,
+                                  DEFAULT_DELAY, DEFAULT_TIMEOUT, DEFAULT_BACKOFF,
+                                  TRANSIENT_SSL_ERROR, TransientSSLError)
+
 
 __all__ = [
     'find',
@@ -39,28 +38,8 @@ __all__ = [
     'reverse_dict',
     'lowercase_keys',
     'get_secure_random_string',
-    'retry',
-
     'ReprMixin'
 ]
-
-# Error message which indicates a transient SSL error upon which request
-# can be retried
-TRANSIENT_SSL_ERROR = 'The read operation timed out'
-
-
-class TransientSSLError(ssl.SSLError):
-    """Represent transient SSL errors, e.g. timeouts"""
-    pass
-
-
-# Constants used by the ``retry`` decorator
-DEFAULT_TIMEOUT = 30  # default retry timeout
-DEFAULT_DELAY = 1  # default sleep delay used in each iterator
-DEFAULT_BACKOFF = 1  # retry backup multiplier
-RETRY_EXCEPTIONS = (RateLimitReachedError, socket.error, socket.gaierror,
-                    httplib.NotConnected, httplib.ImproperConnectionState,
-                    TransientSSLError)
 
 
 def find(l, predicate):
@@ -72,6 +51,9 @@ def find(l, predicate):
 # been moved to "libcloud.common.providers" module
 get_driver = _get_driver
 set_driver = _set_driver
+# Note: This is an alias for backward-compatibility for a function which has been
+# moved to "libcloud.util.retry" module
+retry = Retry
 
 
 def merge_valid_keys(params, valid_keys, extra):
@@ -274,64 +256,4 @@ class ReprMixin(object):
         return str(self.__repr__())
 
 
-def retry(retry_exceptions=RETRY_EXCEPTIONS, retry_delay=DEFAULT_DELAY,
-          timeout=DEFAULT_TIMEOUT, backoff=DEFAULT_BACKOFF):
-    """
-    Retry decorator that helps to handle common transient exceptions.
 
-    :param retry_exceptions: types of exceptions to retry on.
-    :param retry_delay: retry delay between the attempts.
-    :param timeout: maximum time to wait.
-    :param backoff: multiplier added to delay between attempts.
-
-    :Example:
-
-    retry_request = retry(timeout=1, retry_delay=1, backoff=1)
-    retry_request(self.connection.request)()
-    """
-    if retry_exceptions is None:
-        retry_exceptions = RETRY_EXCEPTIONS
-    if retry_delay is None:
-        retry_delay = DEFAULT_DELAY
-    if timeout is None:
-        timeout = DEFAULT_TIMEOUT
-    if backoff is None:
-        backoff = DEFAULT_BACKOFF
-
-    timeout = max(timeout, 0)
-
-    def transform_ssl_error(func, *args, **kwargs):
-        try:
-            return func(*args, **kwargs)
-        except ssl.SSLError as exc:
-            if TRANSIENT_SSL_ERROR in str(exc):
-                raise TransientSSLError(*exc.args)
-
-            raise exc
-
-    def decorator(func):
-        @wraps(func)
-        def retry_loop(*args, **kwargs):
-            current_delay = retry_delay
-            end = datetime.now() + timedelta(seconds=timeout)
-
-            while True:
-                try:
-                    return transform_ssl_error(func, *args, **kwargs)
-                except retry_exceptions as exc:
-                    if isinstance(exc, RateLimitReachedError):
-                        time.sleep(exc.retry_after)
-
-                        # Reset retries if we're told to wait due to rate
-                        # limiting
-                        current_delay = retry_delay
-                        end = datetime.now() + timedelta(
-                            seconds=exc.retry_after + timeout)
-                    elif datetime.now() >= end:
-                        raise
-                    else:
-                        time.sleep(current_delay)
-                        current_delay *= backoff
-
-        return retry_loop
-    return decorator
